ОСВЕЖИ МЕНЯ ДАЙ МНЕ СВЕЖЕСТЬ
ДЕЗОДАРАНТ ИСТРАТЬ НА МОЮ ОДЕЖДУ

+++ lua/nadmin/modules/sh_playtime.lua
@@ -19,11 +19,12 @@ if CLIENT or SERVER then
 end
 
 if SERVER then
-	if nAdminDBFail then return end
+	if not nAdminDB then return end
 	local meta = FindMetaTable'Player'
 
 	function meta:SetPTime(TIME, setnwint)
 		if not IsValid(self) then return end
+		if not nAdminDB then return end
 		local ACID = self:AccountID()
 		local Q = nAdminDB:query("REPLACE INTO nAdmin_time (infoid, time) VALUES (" .. SQLStr(ACID) .. ", " .. SQLStr(TIME) .. ")")
 		function Q:onError(err)
@@ -48,6 +49,7 @@ if SERVER then
 
 	function meta:GetPTime(func)
 		if not IsValid(self) then return end
+		if not nAdminDB then return end
 		local ACID = self:AccountID()
 		local Q = nAdminDB:query("SELECT time FROM nAdmin_time WHERE infoid = " .. SQLStr(ACID) .. " LIMIT 1")
 		function Q:onError(err)
@@ -65,6 +67,7 @@ if SERVER then
 
 	function meta:RemovePTime()
 		if not IsValid(self) then return end
+		if not nAdminDB then return end
 		local ACID = self:AccountID()
 		local Q = nAdminDB:query("DELETE FROM nAdmin_time WHERE infoid = " .. SQLStr(ACID))
 		function Q:onError(err)
@@ -97,14 +100,16 @@ if SERVER then
 	nAdmin.AddCommand("restoretime", false, function(ply)
 		timer.Simple(1, function()
 			--local query = sql.QueryRow("SELECT totaltime FROM utime WHERE player = " .. ply:UniqueID() .. ";")
+			if not nAdminDB then return end
 			local ACID = ply:AccountID()
 			local Q = nAdminDB:query("SELECT * FROM `ulxtime` WHERE `player` = " .. ply:UniqueID())
 			function Q:onError(err)
 				nAdmin.Print("Запрос выдал ошибку: " .. err)
 			end
 			function Q:onSuccess(data)
-				if data and data[1] then
+				if data and data[1] and IsValid(ply) then
 					ply:GetPTime(function(a)
+						if not IsValid(ply) then return end
 						if a > data[1].totaltime then
 							ply:ChatPrint("В базе данных число меньше чем на сервере. Отмена восстановлению часов.")
 							return
@@ -112,6 +117,7 @@ if SERVER then
 						ply:SetPTime(data[1].totaltime + (10 * 60 * 60), true)
 						ply:ChatPrint("Восстановлено!")
 						timer.Simple(3, function()
+							if not IsValid(ply) then return end
 							local a = nAdminDB:query("DELETE FROM `ulxtime` WHERE `player` = " .. ply:UniqueID() .. ";")
 							a:start()
 						end)
